First git commit: git fork of srdy-drdy library from https://opencores.org/projects/srdydrdy_lib
The srdy-drdy library provides a group of components all built around a common data-transfer protocol. This protocol is used in datapath applications and provides bidirectional flow control.

Components in the library provide basic timing closure, clock domain crossing, basic and advanced buffering, and some arbitration and specialized components. The components in the library have been used in multiple successful tape-outs and FPGA designs.

This is the work of Guy Hutchinson, Revision 19 downloaded from
opencores.org 08-2023. The revision is listed with an age of >12 years,
and is considered stable.

The main purpose of this git repository is to enable inclusion in other
git projects as a git submodule.